added argument chaining only for adding todo items for now. doesn't compile and test
