+++ src/functions/mod.rs
@@ -1,2 +1,5 @@
-pub mod commands;
+pub mod done_commands;
+pub mod general_commands;
+pub mod todo_commands;
+
 mod helpers;
